// src/components/Reservation.d.ts
export declare enum Reservation {
    Audio = "audio",
    Camera = "camera",
    Children = "children",
    Dataset = "dataset",
    Filter = "filter",
    Html = "html",
    Image = "image",
    Particle = "particle",
    Physics = "physics",
    Rect = "rect",
    Text = "text",
    Transform = "transform",
    Video = "video"
}